impl Error for CommandError
Mostly Display, but `source` will give the underlying io::Error if
spawning fails.

I'm currently not printing the `io:Error` in the `SpawnFailed` case
of `Display`, depending on the user to print from `source`.  I'm not
sure that's correct.

And I'm not sure tha `SpawnFailed` is always an accurate description.
Is there any other way of getting an `io::Error`?