struct CommandError, enum CommandStatusError
Make (the newly-renamed) `CommandError` struct the default error type
for `stringent`.

It's a struct that's capable of holding `stdout` and `stderr`, while
CommandStatusError is the `enum` that records the type of error
(`SpawnFailed`, `ExitCode`, or `Signal`) and is the type of the `err`
field of `CommandError`. We implement `From<CommandStatusError> for
CommandError`, so a routine with a mix `.status().verify()` and
`.output.verify()` calls, for instance, can have a `Result<Foo,
CommandError>`.